Finish adding JPA entity objects.

+++ API/src/main/java/com/controllers/AccountController.java
@@ -48,7 +48,7 @@ public class AccountController {
 
     @ResponseBody
     @RequestMapping(path = "/id/{id}", method = RequestMethod.GET, produces = "application/json")
-    public Account getAccountRequestHandler(@PathVariable final Long id) {
+    public Account getAccountRequestHandler(@PathVariable final Integer id) {
         return accountRequestHandler.getUserById(id);
     }
